Add spec for making modifications to the repository
This spec file has a few tests for checking that an adapter modifies a
repository as Grit would, when using the form of the `Index#commit` call
used by `Gollum::Committer`.

In particular, one of the specs surfaces gollum/rugged_adapter#11
(where the name of the committer was not the same as the author's).